requirements.txt
streamlit
google-generativeai
python-dotenv
Pillow
opencv-python
streamlit-mermaid
fpdf2
